관리자 고유 기능 구현

+++ jwt/menus/routes.py
@@ -5,11 +5,14 @@ from typing import List
 from core.db import get_db
 from menus.schemas import *
 from menus.service import *
+from users.service import get_user_level
 
 router = APIRouter(prefix="/menu", tags=["menu"])
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
-def create_menu_item(menu: MenuCreateSchema, db: Session = Depends(get_db)):
+def create_menu_item(menu: MenuCreateSchema, db: Session = Depends(get_db), user_level: str = Depends(get_user_level)):
+    if user_level != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="관리자 권한이 필요합니다.")
     create_menu(db, menu)
     return {"message": "메뉴 등록이 완료되었습니다."}
 
@@ -21,11 +24,16 @@ def list_menu_items(skip: int = 0, limit: int = 100, db: Session = Depends(get_d
 def read_menu(menu_name: str, db: Session = Depends(get_db)):
     return get_menu_by_name(db, menu_name)
 
-@router.patch("/{menu_name}", response_model=MenuResponseSchema)
-def update_menu_item(menu_name: str, menu_data: MenuUpdateSchema, db: Session = Depends(get_db)):
+@router.patch("/{menu_name}")
+def update_menu_item(menu_name: str, menu_data: MenuUpdateSchema, db: Session = Depends(get_db), user_level: str = Depends(get_user_level)):
+    if user_level != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="관리자 권한이 필요합니다.")
+    update_menu(db, menu_name, menu_data)
     return {"message": "업데이트가 완료되었습니다."}
 
-@router.delete("/{menu_name}", status_code=status.HTTP_200_OK, response_model=dict)
-def delete_menu_item(menu_name: str, db: Session = Depends(get_db)):
+@router.delete("/{menu_name}", status_code=status.HTTP_200_OK)
+def delete_menu_item(menu_name: str, db: Session = Depends(get_db), user_level: str = Depends(get_user_level)):
+    if user_level != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="관리자 권한이 필요합니다.")
     delete_menu(db, menu_name)
     return {"message": "삭제가 완료되었습니다."}
